finished logs and files modified date check

+++ src/main/java/com/company/file/manager/FileManager.java
@@ -1,26 +1,28 @@
 package com.company.file.manager;
 
 import com.company.file.parser.XmlParser;
-import com.company.file.storage.DataStorage;
 import org.apache.log4j.Level;
 
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+
 /**
  * Created by Rostyslav.Pash on 22-Feb-16.
- *
+ * <p/>
  * Class to copy files from srcDir to targetDir
  */
 
 //C:\!Work\copy Test\dir1
 //C:\!Work\copy Test\dir2
 
-public class FileManager extends BaseSettings{
+public class FileManager extends BaseSettings {
     /**
      * Constructor to set path's set in string parameters
+     *
      * @param srcDir
      * @param targetDir
      */
@@ -45,8 +47,8 @@ public class FileManager extends BaseSettings{
         }
         dataStorage = xmlParser.getDataStorage();
         for (int i = 0; i < dataStorage.getSourcePath().size(); i++) {
-                srcDir.add(new File(dataStorage.getSourcePath().get(i)));
-                targetDir.add(new File(dataStorage.getDestPath().get(i)));
+            srcDir.add(new File(dataStorage.getSourcePath().get(i)));
+            targetDir.add(new File(dataStorage.getDestPath().get(i)));
         }
         this.filesList = getListOfFiles();
         Message.log(getProgramCurrentDirectory());
@@ -54,6 +56,7 @@ public class FileManager extends BaseSettings{
 
     /**
      * Method that returns current working directory
+     *
      * @return String
      */
 
@@ -61,26 +64,34 @@ public class FileManager extends BaseSettings{
         return Paths.get(".").toAbsolutePath().normalize().toString() + "\\";
     }
 
+    public void jobDO() {
+        copyAllFiles();
+        createAllSqlFile();
+        runScript();
+    }
+
     /**
      * Method that runs cmd file in current dir
      */
     public void runScript() {
-        Message.log("Running bat files");
-        if (dataStorage.getScriptsToExecute().size() != 0)
-        for (String scriptPath:
-            dataStorage.getScriptsToExecute()) {
-            try {
-                // Jenkins
-                //Runtime.getRuntime().exec("cmd.exe /c " + getProgramCurrentDirectory() + scriptPath);
-                Runtime.getRuntime().exec("cmd.exe /c start " + scriptPath);
-                Message.log("Script: " + scriptPath + " executed!");
-            } catch (IOException io) {
-                Message.log("Bat file was not executed:");
-                Message.log(io.getStackTrace().toString(), Level.ERROR);
-            }
-        }
-
-        Message.log("Running bat files finished");
+        if (areFilesDifferent) {
+            Message.log("****************** Changes in files were made, running bat files *********");
+            if (dataStorage.getScriptsToExecute().size() != 0)
+                for (String scriptPath :
+                        dataStorage.getScriptsToExecute()) {
+                    try {
+                        // Jenkins
+                        //Runtime.getRuntime().exec("cmd.exe /c " + getProgramCurrentDirectory() + scriptPath);
+                        Runtime.getRuntime().exec("cmd.exe /c start " + scriptPath);
+                        Message.log("Script: " + scriptPath + " executed!");
+                    } catch (IOException io) {
+                        Message.log("Bat file was not executed:");
+                        Message.log(io.getStackTrace().toString(), Level.ERROR);
+                    }
+                }
+            Message.log("**************** Running bat files finished ******************************");
+        } else
+            Message.log("**************** No updates in *.sql files, scripts were not executed! ***");
     }
 
     /**
@@ -98,6 +109,7 @@ public class FileManager extends BaseSettings{
 
     /**
      * Methods prints files in dir set in parameter
+     *
      * @param filePath
      */
     public void printListOfFiles(String filePath) {
@@ -132,26 +144,30 @@ public class FileManager extends BaseSettings{
 
     }
 
-    public void createAllSqlFile() throws IOException {
-        for (File targetDir :
-                this.targetDir) {
-            File allSqlFile = new File(targetDir.toString() + "\\All.sql");
-            if (allSqlFile.createNewFile()) {
-                Message.log("File All.sql created in " + targetDir.getPath() + "\\");
-            } else {
-                Message.log("File All.sql already exist " + targetDir.getPath() + "\\");
-            }
-            // Add all filenames to All.sql
-            //FileUtils.writeStringToFile(file, "String to append", true);
-            try(PrintWriter out = new PrintWriter(new BufferedWriter(new FileWriter(targetDir.toString() + "\\All.sql", false)))) {
-                for (File file :
-                        targetDir.listFiles()) {
+    public void createAllSqlFile() {
+        try {
+            for (File targetDir :
+                    this.targetDir) {
+                File allSqlFile = new File(targetDir.toString() + "\\All.sql");
+                if (allSqlFile.createNewFile()) {
+                    Message.log("File All.sql created in " + targetDir.getPath() + "\\");
+                } else {
+                    Message.log("File All.sql already exist " + targetDir.getPath() + "\\");
+                }
+                // Add all filenames to All.sql
+                //FileUtils.writeStringToFile(file, "String to append", true);
+                try (PrintWriter out = new PrintWriter(new BufferedWriter(new FileWriter(targetDir.toString() + "\\All.sql", false)))) {
+                    for (File file :
+                            targetDir.listFiles()) {
                         if (file.getName().endsWith(".sql") && !file.getName().equalsIgnoreCase("All.sql"))
                             out.println("START " + file.getAbsolutePath());
+                    }
+                } catch (IOException e) {
+                    Message.log(e.toString(), Level.ERROR);
                 }
-            }catch (IOException e) {
-                Message.log(e.toString(), Level.ERROR);
             }
+        } catch (IOException e) {
+            Message.log(e.toString(), Level.ERROR);
         }
     }
 
@@ -162,27 +178,53 @@ public class FileManager extends BaseSettings{
         for (int i = 0; i < filesList.size(); i++) {
             for (File file : filesList.get(i)) {
                 try {
-                    if (!targetDir.get(i).isDirectory() || !targetDir.get(i).exists()){
-                        Path targetPathToFile = Paths.get(targetDir.get(i).toString());
-                        Files.createDirectories(targetPathToFile);
-                    }
+                    clearAllRemovedFiles();
                     if (!file.getName().equalsIgnoreCase("all.sql") && !file.isDirectory() && file.getName().endsWith(".sql")) {
-                        InputStream in = new FileInputStream(file);
-                        OutputStream out = new FileOutputStream(targetDir.get(i) + "/" + file.getName());
-                        byte[] buf = new byte[1024];
-                        int len;
-                        while ((len = in.read(buf)) > 0) {
-                            out.write(buf, 0, len);
-                        }
-                        in.close();
-                        out.close();
-                        Message.log(file.getName() + " file copied successfully to " + targetDir.get(i).getPath());
+                        if (!targetDir.get(i).exists() || isFileChanged(file.getPath(), targetDir.get(i) + "/" + file.getName())) {
+                            if (!targetDir.get(i).isDirectory() || !targetDir.get(i).exists()) {
+                                Path targetPathToFile = Paths.get(targetDir.get(i).toString());
+                                Files.createDirectories(targetPathToFile);
+                            }
+                            InputStream in = new FileInputStream(file);
+                            OutputStream out = new FileOutputStream(targetDir.get(i) + "/" + file.getName());
+                            byte[] buf = new byte[1024];
+                            int len;
+                            while ((len = in.read(buf)) > 0) {
+                                out.write(buf, 0, len);
+                            }
+                            in.close();
+                            out.close();
+                            areFilesDifferent = true;
+                            Message.log(file.getName() + " file copied successfully to " + targetDir.get(i).getPath());
+                        } else
+                            Message.log(file.getName() + " file NOT copied to " + targetDir.get(i).getPath());
                     } else
                         Message.log(file.getName() + " file NOT copied to " + targetDir.get(i).getPath());
                 } catch (IOException e) {
                     e.printStackTrace();
                 }
             }
+        }
+    }
+
+    private void clearAllRemovedFiles() {
+        for (int i = 0; i < srcDir.size(); i++) {
+            
+        }
+        
+    }
+
+    public boolean isFileChanged(String sourceFilePath, String targetFilePath) {
+        File sourceFile = new File(sourceFilePath);
+        File targetFile = new File(targetFilePath);
+        SimpleDateFormat sdf = new SimpleDateFormat("MM/dd/yyyy HH:mm:ss");
+        if (sourceFile.lastModified() > targetFile.lastModified()) {
+            Message.log("New version: " + sourceFilePath + " modified: " + sdf.format(sourceFile.lastModified()));
+            Message.log("Old version: " + targetFilePath + " modified: " + sdf.format(targetFile.lastModified()));
+            return true;
+        } else {
+            Message.log(sourceFilePath + " is up to date");
+            return false;
         }
     }
 }
